tidy up sysinfo() on BSDs
On OpenBSD and NetBSD, the kernel build date helps indicate
patchlevel, so add that to kernel=. This displaces uname -m,
so move that to the OS.

For shell versions, the version of ksh and tcsh aren't terribly
interesting---notably, pdksh has had the same version string
for more than two decades. Instead, only print the version string
if the shell is bash or zsh, both of which regularly add features.