free() display hugepages separately
As far as the kernel is concerned, hugepages are a different pool of memory than that available to "regular" programs. As such, print hugepage totals separately, and subtract hugepage allocation from the "total" amount of (regular) memory available in order to give a more accurate accounting of the size of each pool.