further cleanup
According to the manual, `set -o local_options` is only useful if you're using `set -o $optname` within a function, which neither zshrc nor kshrc do. I think I didn't read the manual thoroughly enough at one point and had thought this was used to set the scope of variables as local to the function by default, which would be very useful but is not what this does. Additionally, the manual suggests `local_options` should be paired with `local_traps`, which was never set here.