pdf.js: use system fonts
Ordinarily PDFs can request loading fonts which aren't installed on the system, which can result in exceptionally poor kerning, at times leaving documents unreadable.

Disable all of this madness and fall back to the system-wide fonts specified in fontconfig. With this we lose some fidelity to the original, but at least the PDF is readable this way.

In the PDFs tested, falling back to system fonts results in a rendering more or less indistinguishable from that of chromium and mupdf, so we must be doing something right.