set git-config advice.skippedCherryPicks = false
This tunable warns during `git pull` if a commit contained in the local worktree has already been cherrypicked upstream, and the local version of the commit was erased as part of the `git pull --rebase`.

Re-ordering commits is pretty common in Gerrit Code Review, and ought to be more common in general---see e.g. jujutsu, https://martinvonz.github.io/jj/.

Without this patch, pulling in a cherrypicked commit prints

```
$ git pull
From ssh://<url>
   88ee45b..394e22e  main       -> origin/main
   warning: skipped previously applied commit 4a915d5
   hint: use --reapply-cherry-picks to include skipped commits
   hint: Disable this message with "git config advice.skippedCherryPicks false"
   Successfully rebased and updated refs/heads/main.
```

In the interest of Don't Print Lots Of Brightly Coloured Text When Everything Is Fine Actually, squelch the warning as suggested in the hint.

Link: https://review.daemonica.net/id/I89e78562e6ce8303ddd23d05bb1fdda69ae365f8
Reviewed-on: https://review.daemonica.net/c/dgoerger/dotfiles/+/122
Reviewed-by: David Goerger <dgoerger@daemonica.net>