add UV index info to current weather conditions
Extend the current weather conditions section to include the UV index value as forecast by the US Environmental Protection Agency.

While here, shorten output outside of the morning report and midday check-ins, and implement a one-level fallback to the next-nearest National Weather Service observation station if the nearest one is offline at the time of query. If both stations are offline, assert that the current temperature is absolute zero (-273 Kelvin) so that it will be obvious we're operating with bad data.

```
AQI: 77 (Moderate, OZONE); UV: 4
Feels like: 29ºC (27ºC, 79% humidity)
```

Change-Id: I83e90eca3091fff925d432a73274fe11b54e82a6
Reviewed-on: https://review.daemonica.net/c/dgoerger/dotfiles/+/142
Reviewed-by: David Goerger <dgoerger@daemonica.net>