import helper script for uploading changes to gerrit
Changes are normally uploaded to Gerrit Code Review by pushing to the special `refs/for/` namespace. For example,

```
$ git push origin HEAD:refs/for/main
```

This introduces a small bit of cognitive overhead:

  1. this is a lot of typing
  2. one must remember whether the target branch for a given repo is `main` or `master`
  3. some checkouts have more than one remote url configured
  4. not all checkouts are configured to use gerrit at all

This helper script is far from perfect, but serves as a starting point to help automate some of the toil by introducing a new `git upload [target_branch]` command line alias.

Change-Id: I88cd4d4a711f9d898d2eb0547cb16c17e0c6f69f
Reviewed-on: https://review.daemonica.net/c/dgoerger/dotfiles/+/183
Reviewed-by: David Goerger <dgoerger@daemonica.net>