gerrit_upload.sh: fix double-negative logic error, improve comments
Double negatives are hard, so let's add a comment to explain what's happening. While here, swap out a `pre-commit run --all-files` (overly broad, something we want in CI) with a more targeted run such that we only trigger pre-commit against the files modified in the latest commit(s) not yet on origin.

Change-Id: Id6d1891815373180acb7bad10c9cb6db331115aa
Reviewed-on: https://review.daemonica.net/c/dgoerger/dotfiles/+/285
Reviewed-by: David Goerger <dgoerger@daemonica.net>