enable git diff --color-moved
The "color-moved" option can be used during major refactorings or code moves to highlight and accentuate whether lines have merely been moved around, versus added or removed. Without this option, lines are colourised on the binary of whether the line has been added (+) or removed (-); this option adds a third category of "moved without semantic change."

It's unfortunate and curious that no code review tool outside of reviewboard appears to take advantage of this option yet.

Change-Id: Icebf3d26759e52c29be1a386ed1b1d20bcc05550
Reviewed-on: https://review.daemonica.net/c/dgoerger/dotfiles/+/302
Reviewed-by: David Goerger <dgoerger@daemonica.net>