enable gitsigns + nvim-tree, bound to ctrl+a
This provides a quick way to enable a vscode-like tree-view file browsing experience with gitsigns integration (to know which lines have changed), and then disable it again so that it gets out of the way and lets an editor be an editor. One of my main pain points in using the sidebar buffers is that copy/paste in the terminal also ends up copying the sidebar text (line numbers, etc), which is pretty much never what I want. This sets the stage for a quick on/off "IDE vs text editor" mode toggle to which I might add other bells and whistles later.

Change-Id: I3af98fa9a878a600dd9bc84b5942f08b84539a7f
Reviewed-on: https://review.daemonica.net/c/dgoerger/dotfiles/+/444
Reviewed-by: David Goerger <dgoerger@daemonica.net>