add more filetypes to nvim-treesitter config
Neovim bundles syntax for most of these filetypes, but we can pull in
upstream fixes sooner with :TSUpdate. This is a test sentence which
should be deleted. Ah nice!

Prompted because the gitcommit format bundled with 0.10.1 was using high
cpu on my box, fixed by using the plugin version instead of the bundled
one.

Change-Id: Ifc82708608b2db3f9bae3276d1e734324e09594f
Reviewed-on: https://review.daemonica.net/c/dgoerger/dotfiles/+/463
Reviewed-by: David Goerger <dgoerger@daemonica.net>