document common package installs
This list is not complete by any means, but is fairly representative of
the packages installed across all of my devices. This does not include
packages installed on a single device, e.g. my VPS has a few extra
packages for server software not found on desktops or other VMs.

Change-Id: I6c795ae769ae1695813a0527d6b00ab456ac2393
Reviewed-on: https://review.daemonica.net/c/dgoerger/dotfiles/+/515
Reviewed-by: David Goerger <dgoerger@daemonica.net>