workaround for gnu printf(1) insanity
bash(1) has a built-in printf command which accepts an optional -v flag:

> The -v option causes the output to be assigned to the variable var
> rather than being printed to the standard output.

AT&T ksh(1) does not have a built-in, so it falls through to printf(1)
from coreutils, which supports flags for --help and --version. The help
flag essentially prints the full manual to stdout, and the version flag
prints the coreutils version. I'm not sure anyone cares about the
coreutils version for something like printf(1), but sure.

OpenBSD ksh(1) doesn't have a built-in either, so it falls through to
printf(1), which on OpenBSD does not support command line flags.

zsh(1) has a built-in printf function which behaves like OpenBSD's
printf(1), i.e. does not accept flags even when running on GNU/Linux.

Now suppose we have a shell script which has something like

```
printf "Choose one:\n"
printf "- option 1\n"
printf "- option 2\n"
```

... the script will fail under both bash and ksh-on-GNU/Linux, but will
succeed under zsh and ksh-on-busybox/Linux and ksh-on-OpenBSD.. because
in the former two scenarios it interprets the quoted hyphen as a command
flag! But the script will fail differently under the two shells, because
the bash built-in and coreutils variant accept different options (short
opts and long opts, respectfully).

As a workaround for this non-portable mayhem, use the asterisk as a
bullet marker instead of hyphen.

Change-Id: I91a6b9b9f03bfe12f1f662e5ad12dc40c90f25d9
Reviewed-on: https://review.daemonica.net/c/dgoerger/dotfiles/+/547
Reviewed-by: David Goerger <dgoerger@daemonica.net>