make pomodoro() compatible with non-BSD systems
The leave(1) command is only distributed out of the box on BSD-based
operating systems. Although the code is highly portable and also works
on Linux, it's not available by default, and we can reimplement this
simply using a combination of the sleep(1) and tput(1) commands, along
with a healthy dose of bourne shell subterfuge.

This introduces a new dependency on tput(1), but the tput(1) command is
installed out of the box on many more systems than leave(1) is, so this
is probably fine. Terminals not supporting bell were already broken as
that's what leave(1) uses under the hood.

While here, reintroduce support for the ${message} flag, which had been
silently ignored ever since we dropped support for libnotify popup
toaster notifications on desktops.

Change-Id: Ic93d8fec56d36b68e7f79a94681d8427948cc6c1
Reviewed-on: https://review.daemonica.net/c/dgoerger/dotfiles/+/586
Reviewed-by: David Goerger <dgoerger@daemonica.net>