add helix editor config
Helix is a neovim-like editor written in rust, which compared with
neovim is at greater liberty to break with historical 'vi' behaviour.

Similar to neovim helix uses the tree-sitter parser, however it's
statically linked (rust-to-rust) instead of dynamically linked as a
system library.

Two major changes introduced by Helix relative to Neovim are

  - kakoune-inspired 'selection + verb' syntax
  - enable all available LSPs automatically

I haven't committed to switching off neovim (discarding 2+ decades of
finger muscle memory is not to be done lightly!), but I'll try it for
now and see how things go. :)

Change-Id: Id9d23c2ae575a9e6de08f32886be5e0d2403d5b4
Reviewed-on: https://review.daemonica.net/c/dgoerger/dotfiles/+/845
Reviewed-by: David Goerger <dgoerger@daemonica.net>