prefer helix over neovim where available
The helix editor[1] is a vim-like editor inspired in part by kakoune[2],
written in rust and with tree-sitter and language server support built
in.

Many keybindings are different.[3] For example,

- delete a word
  - vi: `dw`
  - hx: `wd` (select word, delete)

- change a word
  - vi: `cw`
  - hx: `ec` (select to end of word, change)

- delete a character
  - vi: `x`
  - hx: `d`

- copy a line
  - vi: `yy`
  - hx: `Xy` (select line, copy)

- global replace
  - vi: `:%s/word/replacement/g<ret>`
  - hx: `%sword<ret>creplacement<esc>`

(`%s` selects the entire buffer; `s` opens a prompt for regex; `<ret>`
validates the regex and reduces the selection to each match; `c` deletes
the selection contents and enters insert mode; then one enters the
replacement text, and `<esc>` returns us to normal mode.)

- go to first line
  - vi: `gg`
  - hx: `gg`

- go to last line
  - vi: `G`
  - hx: `ge`

- go to line start
  - vi: `0`
  - hx: `gh`

- go to the line's first non-blank character
  - vi: `^`
  - hx: `gs`

- go to line end
  - vi: `$`
  - hx: `hl`

- delete to the end of the line:
  - vi: `D`
  - hx: `vgld`

(`v` enter visual mode; then `gl` select to line end; then delete)

- delete entire line
  - vi: `dd`
  - hx: `xd` (select line, delete)

- jump to matching bracket
  - vi: `%`
  - hx: `mm`

- autocomplete
  - vi: `C-p`
  - hx: `C-x`

- comment lines
  - vi: `gc`
  - hx: `Space-c`

- search for the word under the cursor
  - vi: `*`
  - hx: `A-o*n` (tree-sitter or lsp) or `be*n`

(`A-o`, alt-o, expands selection to the parent syntax node, i.e. the
word; then `*` use the current selection as the search pattern; then `n`
go the next occurence. `b` selects to the beginning of the word, and `e`
to the end, effectively selecting the whole word in situations where
tree-sitter grammar is unavailable.)

[1] https://helix-editor.com

[2] https://kakoune.org

[3] https://docs.helix-editor.com/from-vim.html

Change-Id: I57d32e33e55e07988243c7f4f3b50d559db3426d
Reviewed-on: https://review.daemonica.net/c/dgoerger/dotfiles/+/863
Reviewed-by: David Goerger <dgoerger@daemonica.net>