remove linters from README
Linters are now enforced by pre-commit. The section in the readme
predates automatic enforcement, had not been updated in years, and was
missing quite a few linters---notably all of the python ones. Zap it
from the readme, which was only ever "enforced" by manual runs, to keep
things DRY and better match reality.

Change-Id: I475b4783e01972f33a012cb133aabedfcb124e69
Reviewed-on: https://review.daemonica.net/c/dgoerger/dotfiles/+/1013
Reviewed-by: David Goerger <dgoerger@daemonica.net>