import jj user config
Jujutsu (jj-vcs) is a new way to interact with git-backed repositories,
with user ergonomics modelled on mercurial, and with first-class
support for tracking and resolving merge conflicts and rebases, in the
style of darcs.

This is an initial config for me to start playing around with jujutsu
in my home lab.

Change-Id: Idc74c8c8568935eaee5a229e5bbfb8766a6a636c
Reviewed-on: https://review.daemonica.net/c/dgoerger/dotfiles/+/1046
Reviewed-by: David Goerger <dgoerger@daemonica.net>